add a per-site scatter plot with letters for key sites in `compare_binding`
@Bernadetadad, this adds code to your `compare_binding` notebook in `custom_rules`
that makes a scatter plot of mutations at key sites across different experiments
(eg, KP.3.1.1 vs XBB.1.5 spikes).

This could be a good way to plot shifts in effects at key sites.

Note that since I have not run the pipeline for a while, I am pushing this
code change without actually running it in the pipeline, as that will take a long
time to complete. So before merging this, pull the branch, run it, push that, and
then merge.